extract InMemoryCatalog out of test
